(feat) Implement strategies store
